+++ src/common/pages/auth.tsx
@@ -14,7 +14,8 @@ import {addUser} from "../store/users";
 import {setActiveUser, updateActiveUser} from "../store/active-user";
 
 import {getAccount} from "../api/hive";
-import {hsTokenRenew, usrActivity} from "../api/private-api";
+import {usrActivity} from "../api/private-api";
+import {hsTokenRenew} from "../api/auth-api";
 
 
 interface Props {
